Allowing the User to Add Ingredients to the Shopping List Completed
Section 6: Course Project - Components & Databinding -> 89 Completed
Added Local References to shopping-edit.component.ts, then created the references to pass data to the shopping-list.component.html in the form of an Ingredient Object by creating and emitting an event.
Further, event listener was created in shopping-list.component.html to listen to the passed data and then pass that data to a method to add it to the array which will be used to read data to display on the screen.