Using the Renderer to build a Better Attribute Directive Completed
Section 7: Directives Deep Dive -> 94 Completed
This Approach is BETTER because Angular is not limited to running in the browser as it also works with Service Workers. These are the environments where we might not have access to the DOM. So, if you try to change the DOM as we did in BasicHighlight by directly accessing the native element and the style of the element, then we might get an error in some circumstances.